% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics.R
\name{analytics}
\alias{analytics}
\title{Analyze the clean data frame extracted from twitter website}
\usage{
analytics(input_file)
}
\arguments{
\item{input_file}{A character vector with, at most, one element.}
}
\value{
A tibble including metrics of analytics.
}
\description{
Analyze the clean data frame extracted from twitter website
}
\examples{
x <- "df.csv"
analytics("x")

}
