% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_tweets.R
\name{clean_tweets}
\alias{clean_tweets}
\title{Cleans the text in the tweets and returns as new columns in the dataframe.
The cleaning process includes converting into lower case, removal of
punctuation, hastags and hastag counts}
\usage{
clean_tweets(file_path, tokenization = TRUE, word_count = TRUE)
}
\arguments{
\item{file_path}{character
File path to csv file containing tweets data}

\item{tokenization}{logical
Creates new column containing cleaned tweet word tokens when True
Default is True}

\item{word_count}{logical
Creates new column containing word count of cleaned tweets
Default is True}
}
\value{
df_tweets
Dataframe
Dataframe containing data on cleaned tweets
.
}
\description{
Cleans the text in the tweets and returns as new columns in the dataframe.
The cleaning process includes converting into lower case, removal of
punctuation, hastags and hastag counts
}
\examples{
clean_tweets("output/tweets_response.csv")


}
