% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotting}
\alias{plotting}
\title{Plot the most frequently occurring hash tags}
\usage{
plotting(tweets_df, text)
}
\arguments{
\item{tweets_df}{a data frame or tibble}

\item{text}{column containing text}
}
\value{
hash_plot
A \code{ggplo2} plot
}
\description{
\code{plotting} creates a horizontal bar chart using \code{ggplot2} syntax.
}
